project/api/api/api_framework/requirements.txt
allure-pytest==2.8.31
allure-python-commons==2.8.31
atomicwrites==1.4.0
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cssselect==1.1.0
cssutils==2.3.0
dailyrotatingfilehandler==0.0.5
ddt==1.4.2
enum34==1.1.10
et-xmlfile==1.1.0
execnet==1.9.0
Flask==1.1.2
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonpath==0.82
lxml==4.6.3
MarkupSafe==1.1.1
more-itertools==8.8.0
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient== 2.0.3
namedlist==1.8
numpy==1.20.3
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
pip==20.1.1
pluggy== 0.13.1
premailer==3.9.0
py==1.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==5.4.3
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-rerunfailures==10.0
pytest-xdist==2.2.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml== 0.17.10
ruamel.yaml.clib==0.2.2
selenium==3.141.0
setuptools==47.1.0
six==1.16.0
template==0.7.2
toml==0.10.2
urllib3==1.26.5
wcwidth==0.2.5
Werkzeug==1.0.1
yagmail==0.14.256